backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Criando Api REST usando TDD com fastapi"
authors = ["Walber Vaz da Silva <wvs.walber@gmail.com>"]
readme = "README.md"
packages = [{ include = "api" }]

[tool.poetry.dependencies]
python = "3.11.*"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic = {extras = ["email"], version = "^2.3.0"}
sqlalchemy = "^2.0.21"
pydantic-settings = "^2.0.3"
alembic = "^1.12.0"
faker = "^19.6.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
taskipy = "^1.12.0"
blue = "^0.9.1"
ruff = "^0.0.291"
httpx = "^0.25.0"
isort = "^5.12.0"
pytest-asyncio = "^0.21.1"
factory-boy = "^3.3.0"
freezegun = "^1.2.2"

[tool.ruff]
line-length = 79
exclude = ['.venv', 'migrations']

[tool.isort]
profile = "black"
line_length = 79
extend_skip = ['migrations']

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = '(migrations/)'

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn api.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=api -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
